Refactoring: ScreeningService findScreeningsByMovieTheaterDate method

+++ src/main/java/com/example/theater_proj/movie/dto/response/SingleSreeningDTO.java
@@ -9,7 +9,7 @@ public record SingleSreeningDTO(
         int id,
         LocalDateTime startTime,
         LocalDateTime endTime,
-        long reservableSeatsCount
+        int reservableSeatsCount
 ) {
 
 }
